backend/requirements.txt
Django==5.0.14
djangorestframework==3.15.0
django-cors-headers==4.3.1
mssql-django==1.4
pyodbc==5.1.0
python-dotenv==1.0.1
Pillow==10.3.0  # For image handling